GTC-2950 Remove WarpedVRT usage in OTF lambda
It looks like WarpedVRT was added to deal with the case where not all
layers in an OTF query were on the same grid. But now the layers that
can be used in an OTF query and are sent in the data_environment by the
Data API must all be on the same grid as the primary dataset. So, we can
remove the use of WarpedVRT. I didn't remove the grid argument in many
of these functions, even though it is no longer used.

I also some unused imports and several long-commented-out lines (related
to grids).

We had two tests test_radd_alerts and test_glad_s2_alerts that were
mainly using a 10/100000 grid but references a layer with a 10/40000
grid. So, I created a new DATA_ENVIRONMENT_10M with only 10/100000
layers, and used that instead for those two tests.

I also fixed up requirements-dev.txt, which was missing some recently
added layers that needed to be installed locally. I think in Github
tests, those are otherwise installed by setup.py, but that isn't used if
you just do "pipenv install -r requirements-dev.txt" locally.