Increase default RESULTS_CHECK_TIMEOUT to 90 seconds
This is a timeout that is written into wait_for_results() in
results_store.py, and it limits the total time to do a tiled raster
analysis, even if the timeouts of the individual lambdas is increased.
So, increased this timeout from 30 to 90 seconds, so we can test doing
OTF on larger GADM areas.